store: refactor schema migrations
Each migration will live in its own file under `store/migrations`

All the sql files in that folder will be embedded into the Go binary so `fewsatscli` can be execute the migrations without any external setup.

Everytime a command is executd we will check if there are any migrations to run and if so, apply them.

 Apply database migrations during initialization