cli: Add --verbose flag to increase logging granularity
By default, the CLI only prints user-facing messages. (`fmt.Print`)

To increase the number of logs and their granularity for debugging, the `--verbose` flag can be used.

When the `--verbose` flag is active, the `slog.Logger` package will be configured to the level set in the profile config and extra logs will be printed in the standard output.

By default all logs will be discarded.